fix and move rsc logic
